update activities sections component
