+++ renderer/src/components/ActivitiesSection/ActivitiesSection.tsx
@@ -1,22 +1,20 @@
 import { useEffect, useMemo, useState } from "react";
-import { ClockIcon, ExclamationCircleIcon } from "@heroicons/react/24/outline";
+import { ExclamationCircleIcon } from "@heroicons/react/24/outline";
 import { ActivitiesTable } from "../ActivitiesTable";
 import { ErrorPlaceholder, RenderError } from "@/shared/ErrorPlaceholder";
-import { PlusIcon } from "@heroicons/react/24/solid";
-import { Square2StackIcon } from "@heroicons/react/24/outline";
 import { loadGoogleEventsFromAllUsers } from "@/helpers/utils/google";
 import { getOffice365Events } from "@/helpers/utils/office365";
 import { checkIsToday } from "@/helpers/utils/datetime-ui";
-import { ButtonTransparent } from "@/shared/ButtonTransparent";
-import Popup from "@/shared/Popup/Popup";
-import { useMainStore } from "@/store/mainStore";
-import { shallow } from "zustand/shallow";
-import { ActivitiesSectionProps, PlaceholderProps } from "./types";
+import { ActivitiesSectionProps } from "./types";
 import { XMarkIcon } from "@heroicons/react/24/solid";
 import { validation } from "@/helpers/utils/reports";
 import { IPC_MAIN_CHANNELS } from "../../../../electron-src/helpers/constants";
+import Placeholder from "./Placeholder";
+import TrackTimeButton from './TrackTimeButton';
+import { RELEASES_LINK } from './constants';
+import { LOCAL_STORAGE_VARIABLES, KEY_CODES } from '@/helpers/contstants';
 
-export default function ActivitiesSection({
+const ActivitiesSection = ({
   onEditActivity,
   activities,
   onDeleteActivity,
@@ -24,7 +22,7 @@ export default function ActivitiesSection({
   latestProjAndAct,
   setSelectedDateReport,
   showAsMain,
-}: ActivitiesSectionProps) {
+}: ActivitiesSectionProps) => {
   const [backgroundError, setBackgroundError] = useState("");
   const [events, setEvents] = useState([]);
   const [isLoading, setIsLoading] = useState(false);
@@ -34,20 +32,18 @@ export default function ActivitiesSection({
   });
   const [errorType, setErrorType] = useState<null | "updater">(null);
   const [isErrorShown, setIsErrorShown] = useState<boolean>(true);
-  const RELEASES_LINK =
-    "https://github.com/ukad-group/timetracker-desktop-client/releases";
-  const isShowGoogleEvents = JSON.parse(
-    localStorage.getItem("showGoogleEvents")
+  const isGoogleEventsShown = JSON.parse(
+    localStorage.getItem(LOCAL_STORAGE_VARIABLES.SHOW_GOOGLE_EVENTS)
   );
-  const isShowOffice365Events = JSON.parse(
-    localStorage.getItem("showOffice365Events")
+  const isOffice365EventsShown = JSON.parse(
+    localStorage.getItem(LOCAL_STORAGE_VARIABLES.SHOW_OFFICE_365_EVENTS)
   );
   const nonBreakActivities = useMemo(() => {
     return validation(activities.filter((activity) => !activity.isBreak));
   }, [activities]);
 
-  const ctrlSpaceHandler = (e: KeyboardEvent) => {
-    if ((e.ctrlKey || e.metaKey) && e.code === "Space") {
+  const handleCtrlSpace = (e: KeyboardEvent) => {
+    if ((e.ctrlKey || e.metaKey) && e.code === KEY_CODES.SPACE) {
       onEditActivity("new");
     }
   };
@@ -56,10 +52,10 @@ export default function ActivitiesSection({
     if (checkIsToday(selectedDate)) {
       setIsLoading(true);
 
-      const googleEvents = isShowGoogleEvents
+      const googleEvents = isGoogleEventsShown
         ? await loadGoogleEventsFromAllUsers()
         : [];
-      const office365Events = isShowOffice365Events
+      const office365Events = isOffice365EventsShown
         ? await getOffice365Events()
         : [];
 
@@ -77,22 +73,22 @@ export default function ActivitiesSection({
 
     loadEvents(isAvailable);
 
-    global.ipcRenderer.on("window-focused", () => {
+    global.ipcRenderer.on(IPC_MAIN_CHANNELS.WINDOW_FOCUSED, () => {
       loadEvents(isAvailable);
     });
 
     return () => {
       setIsLoading(false);
       isAvailable.isAvailable = false;
-      global.ipcRenderer.removeAllListeners("window-focused");
+      global.ipcRenderer.removeAllListeners(IPC_MAIN_CHANNELS.WINDOW_FOCUSED);
     };
   }, [selectedDate]);
 
   useEffect(() => {
-    document.addEventListener("keyup", ctrlSpaceHandler);
+    document.addEventListener("keyup", handleCtrlSpace);
     global.ipcRenderer.on(
       IPC_MAIN_CHANNELS.BACKEND_ERROR,
-      (event, errorMessage, data) => {
+      (_, errorMessage, data) => {
         setBackgroundError(errorMessage);
         console.log("Error data ", data);
 
@@ -104,25 +100,25 @@ export default function ActivitiesSection({
     );
 
     global.ipcRenderer.on(
-      "render error",
-      (event, errorTitle, errorMessage, data) => {
+      IPC_MAIN_CHANNELS.RENDER_ERROR,
+      (_, errorTitle, errorMessage, data) => {
         setRenderError({ errorTitle, errorMessage });
         console.log("Error data ", data);
       }
     );
 
     return () => {
-      document.removeEventListener("keyup", ctrlSpaceHandler);
+      document.removeEventListener("keyup", handleCtrlSpace);
       global.ipcRenderer.removeAllListeners(IPC_MAIN_CHANNELS.BACKEND_ERROR);
       global.ipcRenderer.removeAllListeners("render or fetch error");
     };
   }, []);
 
-  const updateDownloadClickHandler = () => {
+  const handleUpdateDownloadClick = () => {
     global.ipcRenderer.send(IPC_MAIN_CHANNELS.REDIRECT, RELEASES_LINK);
   };
 
-  const closeBtnHandler = () => {
+  const handleCloseButton = () => {
     setIsErrorShown(false);
   };
 
@@ -159,19 +155,18 @@ export default function ActivitiesSection({
               {errorType === "updater" && (
                 <button
                   className="text-dark-button-back hover:text-dark-button-hover"
-                  onClick={updateDownloadClickHandler}
+                  onClick={handleUpdateDownloadClick}
                 >
                   You can download the new version manually from the link
                 </button>
               )}
               <XMarkIcon
                 className="w-6 h-6 fill-gray-600 dark:fill-gray-400/70 absolute right-1 top-1 cursor-pointer"
-                onClick={closeBtnHandler}
+                onClick={handleCloseButton}
               />
             </div>
           </div>
         )}
-
         <div className="px-4 py-5 sm:px-6">
           <ActivitiesTable
             onEditActivity={onEditActivity}
@@ -185,117 +180,12 @@ export default function ActivitiesSection({
             nonBreakActivities={nonBreakActivities}
           />
         </div>
-
         <div>
-          <button
-            id="newActivityBtn"
-            className="block w-full px-4 py-4 text-sm font-medium text-center text-blue-500 bg-blue-200 hover:bg-blue-300 sm:rounded-b-lg dark:bg-dark-button-back-gray hover:dark:bg-dark-button-gray-hover dark:text-dark-heading"
-            onClick={() => onEditActivity("new")}
-          >
-            Track more time
-            <span className="block text-blue-500 text-xs dark:text-dark-heading">
-              click or press ctrl + space
-            </span>
-          </button>
+          <TrackTimeButton onEditActivity={onEditActivity} />
         </div>
       </div>
     </div>
   );
 }
 
-function Placeholder({
-  onEditActivity,
-  backgroundError,
-  selectedDate,
-  setSelectedDateReport,
-}: PlaceholderProps) {
-  const [showModal, setShowModal] = useState(false);
-  const [reportsFolder] = useMainStore(
-    (state) => [state.reportsFolder, state.setReportsFolder],
-    shallow
-  );
-
-  const copyLastReport = async () => {
-    const prevDayReport = await global.ipcRenderer.invoke(
-      "app:find-last-report",
-      reportsFolder,
-      selectedDate
-    );
-
-    if (prevDayReport) {
-      global.ipcRenderer.invoke(
-        "app:write-day-report",
-        reportsFolder,
-        selectedDate,
-        prevDayReport
-      );
-
-      setSelectedDateReport(prevDayReport);
-    } else {
-      setShowModal(true);
-    }
-  };
-
-  return (
-    <div className="py-6 text-center">
-      {backgroundError && (
-        <div className="border-t-4  border-red-700 mx-3 mb-6 p-5 shadow-lg text-gray-700 dark:text-slate-400 text-left">
-          <div className="flex justify-start gap-2 w-full text-gray-900 dark:text-dark-heading font-bold">
-            <ExclamationCircleIcon
-              className="w-7 h-7 text-red-700"
-              aria-hidden="true"
-            />
-            <p>Noncritical error</p>
-          </div>
-          <div className="pl-9 pr-8">
-            {backgroundError} Refer to the console for specific error
-            information.
-          </div>
-        </div>
-      )}
-      <ClockIcon
-        className="w-12 h-12 mx-auto text-gray-400"
-        aria-hidden="true"
-      />
-
-      <h3 className="mt-2 text-sm font-medium text-gray-900 dark:text-dark-heading">
-        No tracked time
-      </h3>
-      <p className="mt-1 text-sm text-gray-500 dark:text-slate-400">
-        Get started by tracking some activity
-      </p>
-      <div className="mt-6 mb-2">
-        <button
-          onClick={() => onEditActivity("new")}
-          type="button"
-          className="inline-flex items-center px-4 py-2 text-sm font-medium text-white bg-blue-600 border border-transparent rounded-md shadow-sm hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500   dark:bg-dark-button-back  dark:hover:bg-dark-button-hover"
-        >
-          <PlusIcon className="w-5 h-5 mr-2 -ml-1" aria-hidden="true" />
-          New activity
-        </button>
-        <span className="block text-gray-500 text-xs">
-          or press ctrl + space
-        </span>
-      </div>
-      <ButtonTransparent callback={copyLastReport}>
-        <Square2StackIcon className="w-5 h-5" />
-        Copy last report
-      </ButtonTransparent>
-      {showModal && (
-        <Popup
-          title="Failed to copy last report"
-          description="Either the last report is empty or you haven't written it for too long"
-          left="25%"
-          top="160px"
-          buttons={[
-            {
-              text: "Ok",
-              color: "green",
-              callback: () => setShowModal(false),
-            },
-          ]}
-        />
-      )}
-    </div>
-  );
-}
+export default ActivitiesSection;
