+++ renderer/src/components/ActivitiesSection/constants.ts
@@ -1,0 +1,2 @@
+export const RELEASES_LINK =
+  "https://github.com/ukad-group/timetracker-desktop-client/releases";
