+++ renderer/src/helpers/contstants.tsx
@@ -1,8 +1,8 @@
-import { ConnectionsSection } from "../components/ConnectionsSection";
-import { MainInputSection } from "../components/MainInputSection";
-import { ReportsFolderSection } from "../components/ReportsFolderSection";
-import { ThemeSection } from "../components/ThemeSection";
-import { VersionSection } from "../components/VersionSection";
+import { ConnectionsSection } from "@/components/ConnectionsSection";
+import { MainInputSection } from "@/components/MainInputSection";
+import { ReportsFolderSection } from "@/components/ReportsFolderSection";
+import { ThemeSection } from "@/components/ThemeSection";
+import { VersionSection } from "@/components/VersionSection";
 
 export enum SidebarNavItem {
   Connections = "Connections",
@@ -19,3 +19,14 @@ export const SETTING_SECTIONS = {
   [SidebarNavItem.VersionSelect]: <VersionSection />,
   [SidebarNavItem.MainInputSelect]: <MainInputSection />,
 };
+
+export const LOCAL_STORAGE_VARIABLES = {
+  SHOW_GOOGLE_EVENTS: 'showGoogleEvents',
+  SHOW_OFFICE_365_EVENTS: 'showOffice365Events',
+  // ...Add more local storage variables as needed
+}
+
+export const KEY_CODES = {
+  SPACE: 'Space'
+  // ...Add more key codes as needed
+}
